requirements.txt
trio-amqp
pytz
PyYAML
